move into `SystemSettings`
